Fixed a minor bug based on an older design.
In a previous design of the regex engine it did not have an explicit lexing step, and instead just assumed all tokens would be one character wide. In this design the accept state character was appended at the end of parsing instead of the end of lexing, and I had forgotten to stop accounting for an extra character in the parsing stage.